Create README.md
UI: You will notice that the UI appears at the beginning of the scene, there are various options to click on, however the mouse is not visible making the navigation difficult. I had to follow a separate tutorial with a unique script to link the UI buttons to other scenes. With more time I would have attempted using the ESC key to call the UI again once in the Lunar Scene.  FP Camera: I was able to recreate the FP camera from the previous module, however whenever my player moves over something it adjusts to that height with respect to the landscape and will not come back down. This is because gravity has been turned off on the player. When it was enabled the player would slide back to the lowest point in the landscape.  Animations: The “monsters” you see throughout the scene came with a complex set of premade animations (similar to the chest we used in this module). However I was not able use the same tutorial steps we followed with the chest and apply it to the monster. My best guess it the problem was related to the player not recognizing the monster collider and triggering the animation. I was however able to get the monster to perform an animation on start, and an attack animation whenever the player presses the space bar.